+++ crates/storage_vec/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "agdb_storage_vec"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+agdb_db_error = { path = "../db_error", version = "<=0.1.0" }
+agdb_serialize = { path = "../serialize", version = "<=0.1.0" }
+agdb_storage = { path = "../storage", version = "<=0.1.0" }
+
+[dev-dependencies]
+agdb_test_file = { path = "../test_file", version = "<=0.1.0" }
