+++ crates/storage_vec/tests/storage_vec_push_test.rs
@@ -1,0 +1,24 @@
+use agdb_storage::Storage;
+use agdb_storage::StorageFile;
+use agdb_storage_vec::StorageVec;
+use agdb_test_file::TestFile;
+
+#[test]
+fn push() {
+    let test_file = TestFile::new();
+    let storage = std::rc::Rc::new(std::cell::RefCell::new(
+        StorageFile::try_from(test_file.file_name().clone()).unwrap(),
+    ));
+
+    let mut vec = StorageVec::<i64>::try_from(storage.clone()).unwrap();
+    vec.push(&1).unwrap();
+    vec.push(&3).unwrap();
+    vec.push(&5).unwrap();
+
+    assert_eq!(
+        storage
+            .borrow_mut()
+            .value::<Vec::<i64>>(&vec.storage_index()),
+        Ok(vec![1_i64, 3_i64, 5_i64])
+    );
+}
